ai_friend_system/requirement.txt
aiosqlite==0.19.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
backoff==2.2.1
billiard==4.2.4
build==1.3.0
cachetools==6.2.2
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cmake==4.2.0
coloredlogs==15.0.1
distro==1.9.0
durationpy==0.10
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.124.0
fastapi-limiter==0.1.6
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.12.0
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
kubernetes==34.1.0
lit==18.1.8
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
networkx==3.6
nltk==3.9.2
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
ollama==0.1.6
onnxruntime==1.23.2
openai==1.3.0
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
orjson==3.11.5
overrides==7.7.0
packaging==25.0
passlib==1.7.4
pillow==12.0.0
posthog==5.4.0
prompt_toolkit==3.0.52
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pybase64==1.4.3
pycparser==2.23
pydantic==2.5.0
pydantic_core==2.14.1
pydub==0.25.1
Pygments==2.19.2
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.6
pyttsx3==2.90
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.31.0
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.11.4
sentence-transformers==2.7.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.4.6
SpeechRecognition==3.9.0
SQLAlchemy==2.0.23
sse-starlette==3.0.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.15.2
torch==2.1.2+cpu
tqdm==4.67.1
transformers==4.36.2
triton==2.0.0
typer==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.24.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
webrtcvad==2.0.10
websocket-client==1.9.0
websockets==15.0.1
zipp==3.23.0
